creation de la classe de test et implementation de la classe interne lesnotes @Nested pour regrouper le test de l'ensemble des tests de la note Notes

+++ src/exercice2DesNotes/notation/Notes.java
@@ -9,7 +9,7 @@ public class Notes {
     private  int nombreDeNotes = 0;
     private float noteMaximale = 0;
     public float noteMinimale = 0;
-    public float moyenneDesNotes = 0;
+    public int moyenneDesNotes = 0;
 
     public Notes(int capacite){
         this.capacite = capacite;
